refactor parquet dictionary types to be generic
